Update card aspect ratio

+++ Memory/Views/MemoryGameView.swift
@@ -16,7 +16,7 @@ struct MemoryGameView: View {
         HStack {
             ForEach(memoryGame.cards) { card in
                 CardView(card: card)
-                    .aspectRatio(CGFloat(2.0/3.0), contentMode: .fit)
+                    .aspectRatio(2/3, contentMode: .fit)
                     .onTapGesture {
                         self.memoryGame.choose(card)
                     }
